Added the events app.
Currently the events app has the all the functionality of the events app from aniMango.

More testing on edge cases has to be done, but everything appears to be working at this point.

With the events app I've added martor (A django markdown editor). Using this over TinyMCE from aniMango as markdown is easier to work with.

The events app still needs significant amounts of work done to the styling and layout, but I'll do those at a later point.