Added privacy policy and renamed site_info to miscellaneous
The name for the site_info app has been changed to miscellaneous. On the old aniMango site, the site_info app held the homepage, GDPR, society history and exec pages. In this site, most of these have been moved to their own app. Therefore, I felt that the miscellaneous was a more appropriate name.